- sector contract double to float

+++ SDK/Contracts/Data/SectorContract.cs
@@ -12,13 +12,13 @@ namespace SDK.Models
 
         public string Title { get; set; }
 
-        public double BarrierWidth { get; set; }
+        public float BarrierWidth { get; set; }
 
-        public double BarrierHeight { get; set; }
+        public float BarrierHeight { get; set; }
 
-        public double SectorWidth { get; set; }
+        public float SectorWidth { get; set; }
 
-        public double SectorHeight { get; set; }
+        public float SectorHeight { get; set; }
 
         public long Modified { get; set; }
